[BREAKING] Huge Refactoring (read commit message for changes)
This role is a fork of the keycloak role of nkinder. The role by nkinder
is made for a older version of Keycloak and since then some things have
changed. Also I do not have a use-case for some of the features nkinder's
role provided and therefore I don't know if they are working anymore.
This is why I removed some of the features. Also in this commit the code
is refactored and cleaned up.

Below are the changes I made in this commit. They should be enough for
you to migrate to this (and newer) versions of this role, but I
recommend that you take some time and read the readme.

**Breaking:**
* You need to take care of java yourself. This role no longer depends
on geerlingguy.java
* All features with TLS have been removed. I don't have a test case for
it and I don't now if they are working. For example the self-signed
certificate is now (I think) auto-created at boot time
* The default bind address has been changed to "127.0.0.1" instead of
"0.0.0.0". This is because in the setup the role is made for a
reverseproxy is in front of Keycloak
* firewalld is not configured anymore by this role. You should take care
doing this yourself

**Features/Other stuff**
* A new README \o/
* The tasks are now splitted up in multiple files
* Default Keycloak version is now 12.0.1 (The role is only tested with
the provided default version)
* The admin user is not created each time the installation runs, but
only when `keycloak_create_admin` is set to true. This is because
normally after a update the user should be already in the database and
doesn't need to be recreated
* You can now set the name of the Keycloak systemd service with
`keycloak_service_name`
* JAVA_OPTS had been changed to `"-Xms256m -Xmx1024m"` - take care this
could lead to an instable Keycloak if you have a high load of users
* You can set whether you use a reverseproxy with
`keycloak_behind_reverseproxy`
* The molecule tests have been removed because they are outdated